% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au.firth.r
\name{au.firth}
\alias{au.firth}
\title{Firth AU testing}
\usage{
au.firth(m0, m1, r0, r1, lowthresh = 1e-12)
}
\arguments{
\item{m0}{Number of control subjects}

\item{m1}{Number of case subjects}

\item{r0}{Number of control subjects exposed}

\item{r1}{Number of case subjects exposed}

\item{lowthresh}{A threshold for probabilities below to be considered as zero. Defaults to 1e-12.}
}
\value{
A single AU p-value, computed under the Firth test.
}
\description{
Calculates approximate unconditional Firth test p-value for testing independence in 2x2 case-control tables.
The Firth test requires more computational resources than the tests computed in the au.tests function.
}
\examples{
au.firth(15000, 5000, 1, 0)
}

