% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.tests.r
\name{basic.tests}
\alias{basic.tests}
\title{Basic testing}
\usage{
basic.tests(m0, m1, r0, r1)
}
\arguments{
\item{m0}{Number of control subjects}

\item{m1}{Number of case subjects}

\item{r0}{Number of control subjects exposed}

\item{r1}{Number of case subjects exposed}
}
\value{
A vector of p-values, computed under score, likelihood ratio, Wald, Firth, and Fisher's exact tests.
}
\description{
Calculates standard p-values for testing independence in 2x2 case-control tables.
}
\examples{
basic.tests(15000, 5000, 30, 25)
}

